refactor: move Login component into view
